requirements.txt
requests>=2.25.0
python-dotenv>=0.19.0
pyaudio>=0.2.11
hangul-romanize>=1.1.0
